webscraper/requirements.txt
apscheduler==3.9.1
python-dotenv==0.21.0
beautifulsoup4==4.11.1
selenium==4.6.0
pandas==1.5.2
influxdb-client==1.35.0